vimrc: add my own Retab function
vim's retab will turn any group of spaces into a tab, which is
problematic when spaces are used in comments to align with non-tab
characters, or when spaces are used in strings.

My methods only operate on whitespace before the first non-whitespace
character.

Note:

For other peoples' reference, this will not intelligently handle a case
when spaces are used after leading tabs to align with non-whitespace
characters, and you'll have to watch out for that when using Retab to
fix leading indentation.

The following is an example of the problem, where leading tabs are ">
" and spaces are "."

>   int main()
>   {
>   >   int result = some_function("hello", "world",
>   >   ..........................."how", "are", "you");
>   >   return result;
>   }

When using Retab, it would turn into the following:

>   int main()
>   {
>   >   int result = some_function("hello", "world",
>   >   >   >   >   >   >   >   ..."how", "are", "you");
>   >   return result;
>   }

The problem with the above is that the original spaces were used to
maintain alginment with the non-tab characters from the above line,
independently of tab size, but now that they are tabs, the alignment
depends on tab size.