gvimrc: Use select mode like GUIs do
This is part of my effort to "make my own evim" because there's only a
couple of things about evim that I actually want when using a GUI
version of vim.

This commit makes it so shift + arrow keys (or home/end) will behave
like GUIs do, selecting when holding shift, and unselecting when not.

This commit also avoids requiring selection=exclusive, so visual mode
works like it usually does in vim.