gvimrc: Use selection=exclusive
So if selection=inclusive, the last character of a word won't be
selected when double-clicking, because the switch to select mode will
change selection to exclusive without correcting the right corner of the
selection.

So, instead, I use selection=exclusive "by default" for gvim, and
automatically switch to inclusive for visual mode.