vimrc: Add todo regarding "+ register
If I'm ssh'd to a server, chances are the vim version doesn't have X
clipboard support, and won't have a "+ register (unless the server is
Windows).

When I'm ssh'd to such a server, I'll probably still want to be able to
<c-c> and rely on OSC 52 to set my local clipboard, so I'll need a
setting to use another register in place of the + register.

Otherwise, is there a way to enable the "+ register when built without
clipboard support? In other words, can a plugin provide its own version
of "clipboard support" through the "+ register?