vimrc: Add (commented out) t_SI, t_EI, and t_SR
These options set the control sequencees to set cursor shape. I was
originally playing with them to see if I can get vim (not gvim) to use a
bar cursor during select mode (when selection=exclusive), but I haven't
gotten that to work. Now, I'm keeping this for future reference. It is
commented out because it is not portable, and I'm not even sure I like
the look in my terminals all that much (it is easier for me to spot my
block cursor than my bar cursor).