vim: rework selection logic
So in vim, when selection=exclusive and mouse=a, if you select words
from right to left (using double-click/drag towards the left), the
entire word that you started with is excluded from the selection.

So now I'm trying to use selection=inclusive for selection mode, as
well, but a few things about vim make this weird. For example
ctrl-shift-right will make the cursor jump to the start of the next
word, but the selection includes the first character of the word.

These mappings are my attempt to adjust for including the character at
the cursor position, and make vim feel more like traditional Windows or
GTK GUIs.

A benefit of this new approach is that my vim experience should "match"
my gvim experience when it comes to selections.