vimrc: Use Linux tty control sequences to set cursor shape
Vim visual mode is easier to use with a block cursor, because vim will
likely not get an option to include the character *at* the cursor
position for the highlight applied in visual mode.

See: https://groups.google.com/g/vim_use/c/YBocFskMxSA/m/tlEjhZQOBAAJ

This customization was primarily done to get a block cursor for visual
mode when on a Linux tty, but I figured I'd set the other two shapes
while I was at it, too.