Add ssh ControlPath TODO note
Annoyingly, if you use the Hostname directive to override the address or
specify the externally-accessible DNS name for an internal server, it
affects the value of the %h/%C token. The value used is the final
hostname, which means that values of %C can conflict.

For example, in the following configuration, internal-router.site-a.net
and internal-router.site-b.net have the same value of %C, so if you ssh
to internal-router.site-a.net, and then ssh to
internal-router.site-b.net, you'll unexpectedly be connected to
internal-router.site-a.net.

Host internal-router.site-a.net
	Hostname 192.168.1.1
	ProxyJump edge.site-a.net

Host internal-router.site-b.net
	Hostname 192.168.1.1
	ProxyJump edge.site-b.net

Host *
	ControlPath ~/.ssh/mux-%C