Add wrapper for rcm commands
bin/rcup was arbitrarily choosen as the main name for the wrapper script
because I view rcup as the most important command of rcm. lsrc, mkrc,
and rcdn are symlinks to the rcup wrapper. The rcup wrapper knows the
difference based on the program name (aka argv[0], visible in bash as
the "${BASH_SOURCE[0]}" value).